Skip notifications when OCR mode is "skip file"
* Fixes #232
* Move logic to OcrService
