+++ lib/BackgroundJobs/ProcessFileJob.php
@@ -27,22 +27,10 @@ declare(strict_types=1);
 namespace OCA\WorkflowOcr\BackgroundJobs;
 
 use \OCP\Files\File;
-use OC\User\NoUserException;
-use OCA\WorkflowOcr\Helper\IProcessingFileAccessor;
 use OCA\WorkflowOcr\Model\WorkflowSettings;
-use OCA\WorkflowOcr\Service\IEventService;
 use OCA\WorkflowOcr\Service\INotificationService;
 use OCA\WorkflowOcr\Service\IOcrService;
-use OCA\WorkflowOcr\Wrapper\IFilesystem;
-use OCA\WorkflowOcr\Wrapper\IViewFactory;
 use OCP\AppFramework\Utility\ITimeFactory;
-use OCP\Files\FileInfo;
-use OCP\Files\IRootFolder;
-use OCP\Files\Node;
-use OCP\Files\NotFoundException;
-use OCP\IUser;
-use OCP\IUserManager;
-use OCP\IUserSession;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -52,47 +40,19 @@ use Psr\Log\LoggerInterface;
 class ProcessFileJob extends \OCP\BackgroundJob\QueuedJob {
 	/** @var LoggerInterface */
 	protected $logger;
-	/** @var IRootFolder */
-	private $rootFolder;
 	/** @var IOcrService */
 	private $ocrService;
-	/** @var IEventService */
-	private $eventService;
-	/** @var IViewFactory */
-	private $viewFactory;
-	/** @var IFilesystem */
-	private $filesystem;
-	/** @var IUserManager */
-	private $userManager;
-	/** @var IUserSession */
-	private $userSession;
-	/** @var IProcessingFileAccessor */
-	private $processingFileAccessor;
 	/** @var INotificationService */
 	private $notificationService;
 	
 	public function __construct(
 		LoggerInterface $logger,
-		IRootFolder $rootFolder,
 		IOcrService $ocrService,
-		IEventService $eventService,
-		IViewFactory $viewFactory,
-		IFilesystem $filesystem,
-		IUserManager $userManager,
-		IUserSession $userSession,
-		IProcessingFileAccessor $processingFileAccessor,
 		INotificationService $notificationService,
 		ITimeFactory $timeFactory) {
 		parent::__construct($timeFactory);
 		$this->logger = $logger;
-		$this->rootFolder = $rootFolder;
 		$this->ocrService = $ocrService;
-		$this->eventService = $eventService;
-		$this->viewFactory = $viewFactory;
-		$this->filesystem = $filesystem;
-		$this->userManager = $userManager;
-		$this->userSession = $userSession;
-		$this->processingFileAccessor = $processingFileAccessor;
 		$this->notificationService = $notificationService;
 	}
 	
@@ -104,13 +64,10 @@ class ProcessFileJob extends \OCP\BackgroundJob\QueuedJob {
 
 		try {
 			[$fileId, $uid, $settings] = $this->parseArguments($argument);
-			$this->initUserEnvironment($uid);
-			$this->processFile($fileId, $settings);
+			$this->ocrService->runOcrProcess($fileId, $uid, $settings);
 		} catch (\Throwable $ex) {
 			$this->logger->error($ex->getMessage(), ['exception' => $ex]);
 			$this->notificationService->createErrorNotification($uid, 'An error occured while executing the OCR process ('.$ex->getMessage().'). Please have a look at your servers logfile for more details.');
-		} finally {
-			$this->shutdownUserEnvironment();
 		}
 
 		$this->logger->debug('ENDED -- Run ' . self::class . ' job. Argument: {argument}.', ['argument' => $argument]);
@@ -134,88 +91,5 @@ class ProcessFileJob extends \OCP\BackgroundJob\QueuedJob {
 			$uid,
 			$settings
 		];
-	}
-
-	/**
-	 * @param int $fileId  The id of the file to be processed
-	 * @param WorkflowSettings $settings The settings to be used for processing
-	 */
-	private function processFile(int $fileId, WorkflowSettings $settings) : void {
-		$node = $this->getNode($fileId);
-
-		$ocrFile = $this->ocrService->ocrFile($node, $settings);
-
-		$filePath = $node->getPath();
-		$fileContent = $ocrFile->getFileContent();
-		$originalFileExtension = $node->getExtension();
-		$newFileExtension = $ocrFile->getFileExtension();
-
-		// Only create a new file version if the file OCR result was not empty #130
-		if ($ocrFile->getRecognizedText() !== '') {
-			$newFilePath = $originalFileExtension === $newFileExtension ?
-				$filePath :
-				$filePath . ".pdf";
-
-			$this->createNewFileVersion($newFilePath, $fileContent, $fileId);
-		}
-
-		$this->eventService->textRecognized($ocrFile, $node);
-	}
-
-	private function getNode(int $fileId) : ?Node {
-		/** @var File[] */
-		$nodeArr = $this->rootFolder->getById($fileId);
-		if (count($nodeArr) === 0) {
-			throw new NotFoundException('Could not process file with id \'' . $fileId . '\'. File was not found');
-		}
-		
-		$node = array_shift($nodeArr);
-
-		if (!$node instanceof Node || $node->getType() !== FileInfo::TYPE_FILE) {
-			throw new \InvalidArgumentException('Skipping process for file with id \'' . $fileId . '\'. It is not a file');
-		}
-
-		return $node;
-	}
-
-	/**
-	 * * @param string $userId 	The owners userId of the file to be processed
-	 */
-	private function initUserEnvironment(string $userId) : void {
-		/** @var IUser */
-		$user = $this->userManager->get($userId);
-		if (!$user) {
-			throw new NoUserException("User with uid '$userId' was not found");
-		}
-
-		$this->userSession->setUser($user);
-		$this->filesystem->init($userId, '/' . $userId . '/files');
-	}
-
-	private function shutdownUserEnvironment() : void {
-		$this->userSession->setUser(null);
-	}
-
-	/**
-	 * @param string $filePath		The filepath of the file to write
-	 * @param string $ocrContent	The new filecontent (which was OCR processed)
-	 * @param int $fileId		The id of the file to write. Used for locking.
-	 */
-	private function createNewFileVersion(string $filePath, string $ocrContent, int $fileId) : void {
-		$dirPath = dirname($filePath);
-		$filename = basename($filePath);
-		
-		$this->processingFileAccessor->setCurrentlyProcessedFileId($fileId);
-
-		try {
-			$view = $this->viewFactory->create($dirPath);
-			// Create new file or file-version with OCR-file
-			// This will trigger 'postWrite' event which would normally
-			// add the file to the queue again but this is tackled
-			// by the processingFileAccessor.
-			$view->file_put_contents($filename, $ocrContent);
-		} finally {
-			$this->processingFileAccessor->setCurrentlyProcessedFileId(null);
-		}
 	}
 }
