+++ lib/Migration/Version2404Date20220903071748.php
@@ -89,7 +89,7 @@ class Version2404Date20220903071748 extends SimpleMigrationStep {
 				$workflowSettings = json_decode($row['operation'], true);
 				$foundMapping = false;
 				$newLangArr = [];
-				if (!array_key_exists('languages', $workflowSettings)) {
+				if (!$workflowSettings || !array_key_exists('languages', $workflowSettings)) {
 					continue;
 				}
 				$languagesArr = $workflowSettings['languages'];
